+++ src/Domain/Messaging/Command/ProcessQueueCommand.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace PhpList\Core\Domain\Messaging\Command;
 
-use PhpList\Core\Domain\Messaging\Service\CampaignProcessor;
+use PhpList\Core\Domain\Messaging\Repository\MessageRepository;
+use PhpList\Core\Domain\Messaging\Service\MessageProcessingPreparator;
+use PhpList\Core\Domain\Messaging\Service\Processor\CampaignProcessor;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use PhpList\Core\Domain\Messaging\Repository\MessageRepository;
-use PhpList\Core\Domain\Messaging\Service\MessageProcessingPreparator;
 use Symfony\Component\Lock\LockFactory;
-use Symfony\Component\Console\Attribute\AsCommand;
 use Throwable;
 
 #[AsCommand(
