+++ tests/Unit/Domain/Messaging/Command/ProcessQueueCommandTest.php
@@ -8,8 +8,8 @@ use Exception;
 use PhpList\Core\Domain\Messaging\Command\ProcessQueueCommand;
 use PhpList\Core\Domain\Messaging\Model\Message;
 use PhpList\Core\Domain\Messaging\Repository\MessageRepository;
-use PhpList\Core\Domain\Messaging\Service\CampaignProcessor;
 use PhpList\Core\Domain\Messaging\Service\MessageProcessingPreparator;
+use PhpList\Core\Domain\Messaging\Service\Processor\CampaignProcessor;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Application;
